pyproject.toml
[project]
name = "purple-terminal-agent"
version = "0.2.0"
description = "Purple agent for Terminal Bench 2.0 — Hierarchical Planner + Critic Pre-flight + RAG"
readme = "README.md"
requires-python = ">=3.13"

dependencies = [
    "a2a-sdk[http-server]>=0.3.20",
    "openai>=1.0.0",
    "httpx>=0.28.1",
    "pydantic>=2.12.5",
    "uvicorn>=0.38.0",
    "tenacity>=9.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.9.0",
]

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
